+++ xrpl_types/src/types/transaction/offer_create.rs
@@ -1,11 +1,11 @@
 use crate::serialize::{FieldCode, Serialize, Serializer};
-use crate::{Amount, Transaction, UInt32};
+use crate::{Amount, TransactionCommon, UInt32};
 use enumflags2::{bitflags, BitFlags};
 
 /// An `OfferCreate` transaction <https://xrpl.org/offercreate.html>
 #[derive(Debug, Clone)]
 pub struct OfferCreateTransaction {
-    pub common: Transaction,
+    pub common: TransactionCommon,
     pub flags: BitFlags<OfferCreateFlags>,
     pub expiration: Option<UInt32>,
     pub offer_sequence: Option<UInt32>,
